% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guistatscience.R
\name{guistatscience}
\alias{guistatscience}
\title{Graphical User Interface for statscience package}
\usage{
guistatscience(gui = TRUE)
}
\arguments{
\item{gui}{Logical argument, \code{TRUE} or \code{FALSE}. The default is \code{TRUE}}
}
\value{
\code{guistatscience} presents a GUI for ...
}
\description{
\code{guistatscience} A Graphical User Interface (GUI) for
    the statscience package
}
\examples{
# Loading package
library(MCPtests)
if (interactive()) {
  guistatscience(gui = FALSE)
}

}
